+++ test/security_token.js
@@ -38,7 +38,7 @@ contract('SecurityToken', accounts => {
     // investor Details
     let fromTime = latestTime();
     let toTime = latestTime() + duration.days(100);
-    
+
     let ID_snap;
     const message = "Transaction Should Fail!!";
 
@@ -119,8 +119,11 @@ contract('SecurityToken', accounts => {
         account_delegate = accounts[5];
         account_temp = accounts[8];
         token_owner = account_issuer;
-      
+
         // ----------- POLYMATH NETWORK Configuration ------------
+
+        // Step 0: Deploy the Polytoken Contract
+        I_PolyToken = await PolyToken.new();
 
         // STEP 1: Deploy the ModuleRegistry
 
@@ -134,7 +137,7 @@ contract('SecurityToken', accounts => {
 
         // STEP 2: Deploy the GeneralTransferManagerFactory
 
-        I_GeneralTransferManagerFactory = await GeneralTransferManagerFactory.new({from:account_polymath});
+        I_GeneralTransferManagerFactory = await GeneralTransferManagerFactory.new(I_PolyToken.address, {from:account_polymath});
 
         assert.notEqual(
             I_GeneralTransferManagerFactory.address.valueOf(),
@@ -144,7 +147,7 @@ contract('SecurityToken', accounts => {
 
         // STEP 3: Deploy the GeneralDelegateManagerFactory
 
-        I_GeneralPermissionManagerFactory = await GeneralPermissionManagerFactory.new({from:account_polymath});
+        I_GeneralPermissionManagerFactory = await GeneralPermissionManagerFactory.new(I_PolyToken.address, {from:account_polymath});
 
         assert.notEqual(
             I_GeneralPermissionManagerFactory.address.valueOf(),
@@ -154,7 +157,7 @@ contract('SecurityToken', accounts => {
 
         // STEP 4: Deploy the CappedSTOFactory
 
-        I_CappedSTOFactory = await CappedSTOFactory.new({ from: token_owner });
+        I_CappedSTOFactory = await CappedSTOFactory.new(I_PolyToken.address, { from: token_owner });
 
         assert.notEqual(
             I_CappedSTOFactory.address.valueOf(),
@@ -194,9 +197,6 @@ contract('SecurityToken', accounts => {
             "0x0000000000000000000000000000000000000000",
             "STVersion contract was not deployed",
         );
-
-        // Step ANY: Deploy the Polytoken Contract
-         I_PolyToken = await PolyToken.new();
 
         // Step 8: Deploy the SecurityTokenRegistry
 
@@ -289,17 +289,16 @@ contract('SecurityToken', accounts => {
             startTime = latestTime() + duration.seconds(5000);
             endTime = startTime + duration.days(30);
             let bytesSTO = web3.eth.abi.encodeFunctionCall(functionSignature, [startTime, endTime, cap, rate, fundRaiseType, I_PolyToken.address, account_fundsReceiver]);
-            
-            const tx = await I_SecurityToken.addModule(I_CappedSTOFactory.address, bytesSTO, 0, 0, false, { from: token_owner, gas: 5000000 });
 
-            assert.equal(tx.logs[2].args._type, stoKey, "CappedSTO doesn't get deployed");
+            const tx = await I_SecurityToken.addModule(I_CappedSTOFactory.address, bytesSTO, 0, 0, true, { from: token_owner, gas: 5000000 });
+            assert.equal(tx.logs[3].args._type, stoKey, "CappedSTO doesn't get deployed");
             assert.equal(
-                web3.utils.toAscii(tx.logs[2].args._name)
+                web3.utils.toAscii(tx.logs[3].args._name)
                 .replace(/\u0000/g, ''),
                 "CappedSTO",
                 "CappedSTOFactory module was not added"
             );
-            I_CappedSTO = CappedSTO.at(tx.logs[2].args._module);
+            I_CappedSTO = CappedSTO.at(tx.logs[3].args._module);
         });
     });
 
@@ -308,7 +307,7 @@ contract('SecurityToken', accounts => {
             let moduleData = await I_SecurityToken.getModule.call(stoKey, 0);
             assert.equal(web3.utils.toAscii(moduleData[0]).replace(/\u0000/g, ''), "CappedSTO");
             assert.equal(moduleData[1], I_CappedSTO.address);
-            assert.isFalse(moduleData[2]);
+            assert.isTrue(moduleData[2]);
         });
 
         it("Should fails in removing the module from the securityToken", async() => {
@@ -339,7 +338,7 @@ contract('SecurityToken', accounts => {
             let key = await takeSnapshot();
             let tx = await I_SecurityToken.removeModule(transferManagerKey, 0, { from : token_owner });
             assert.equal(tx.logs[0].args._type, transferManagerKey);
-            assert.equal(tx.logs[0].args._module, I_GeneralTransferManager.address);            
+            assert.equal(tx.logs[0].args._module, I_GeneralTransferManager.address);
             await revertToSnapshot(key);
         });
 
@@ -347,7 +346,7 @@ contract('SecurityToken', accounts => {
             let moduleData = await I_SecurityToken.getModule.call(transferManagerKey, 0);
             assert.equal(web3.utils.toAscii(moduleData[0]).replace(/\u0000/g, ''), "GeneralTransferManager");
             assert.equal(moduleData[1], I_GeneralTransferManager.address);
-            assert.isTrue(moduleData[2]);
+            assert.isFalse(moduleData[2]);
         });
 
         it("Should change the budget of the module", async() => {
@@ -363,7 +362,7 @@ contract('SecurityToken', accounts => {
             it("Should Buy the tokens", async() => {
                 balanceOfReceiver = await web3.eth.getBalance(account_fundsReceiver);
                 // Add the Investor in to the whitelist
-    
+
                 let tx = await I_GeneralTransferManager.modifyWhitelist(
                     account_investor1,
                     fromTime,
@@ -372,9 +371,9 @@ contract('SecurityToken', accounts => {
                         from: account_issuer,
                         gas: 500000
                     });
-    
+
                 assert.equal(tx.logs[0].args._investor, account_investor1, "Failed in adding the investor in whitelist");
-    
+
                 // Jump time
                 await increaseTime(5000);
                 // Fallback transaction
@@ -384,16 +383,16 @@ contract('SecurityToken', accounts => {
                     gas: 210000,
                     value: web3.utils.toWei('1', 'ether')
                     });
-    
+
                 assert.equal(
                     (await I_CappedSTO.fundsRaised.call())
                     .dividedBy(new BigNumber(10).pow(18))
                     .toNumber(),
                     1
                 );
-    
+
                 assert.equal(await I_CappedSTO.getNumberInvestors.call(), 1);
-    
+
                 assert.equal(
                     (await I_SecurityToken.balanceOf(account_investor1))
                     .dividedBy(new BigNumber(10).pow(18))
@@ -412,7 +411,7 @@ contract('SecurityToken', accounts => {
                     ensureException(error);
                 }
                 assert.ok(errorThrown, message);
-            });   
+            });
 
             it("Should fail to provide the permission to the delegate to change the transfer bools", async () => {
                 let errorThrown = false;
@@ -497,7 +496,7 @@ contract('SecurityToken', accounts => {
                         from: account_issuer,
                         gas: 500000
                     });
-    
+
                 assert.equal(tx.logs[0].args._investor, account_investor2, "Failed in adding the investor in whitelist");
 
                 await I_SecurityToken.transfer(account_investor2, (10 *  Math.pow(10, 18)), { from : account_investor1});
@@ -550,16 +549,16 @@ contract('SecurityToken', accounts => {
                     gas: 210000,
                     value: web3.utils.toWei('1', 'ether')
                     });
-    
+
                 assert.equal(
                     (await I_CappedSTO.fundsRaised.call())
                     .dividedBy(new BigNumber(10).pow(18))
                     .toNumber(),
                     2
                 );
-    
+
                 assert.equal(await I_CappedSTO.getNumberInvestors.call(), 2);
-    
+
                 assert.equal(
                     (await I_SecurityToken.balanceOf(account_investor1))
                     .dividedBy(new BigNumber(10).pow(18))
@@ -596,7 +595,7 @@ contract('SecurityToken', accounts => {
                     console.log(`non-whitelist investor is not allowed`);
                     errorThrown = true;
                     ensureException(error);
-                } 
+                }
                 assert.ok(errorThrown, message);
            });